Fix SEEK_END, file offset overflow, and read/write/mkpartition syscall bugs.
Fix SEEK_END seeking twice as far as requested. Centralize lseek handling in
one place and avoid overflow bugs. Inode lseek handlers now only need to
handle SEEK_END with offset 0. Prevent the file offset from ever going below
zero or overflowing.

Character devices are now not seekable, but lseek will pretend they are, yet
always stay at the file offset 0. pread/pwrite on character devices will now
ignore the file offset and call read/write.

This change prevents character devices from being memory mapped, notably
/dev/zero can no longer be memory mapped. None of the current ports seem
to rely on this behavior and will work with just MAP_ANONYMOUS.

Refactor read and write system calls to have a shared return statement for
both seekable and non-seekable IO.

Fix file offset overflow bugs in read and write system calls.

Fix system calls returning EPERM instead of properly returning EBADF when
the file has not been opened in the right mode.

Truncate IO counts and total vector IO length so the IO operation does not
do any IO beyond OFF_MAX. Truncate also total vector IO length for recvmsg
and sendmsg. Fail with EINVAL if total vector IO length exceeds SSIZE_MAX.

Don't stop early if the total IO length is zero, so zero length IO now block
on any locks internal to the inode.

Handle reads at the maximum file offset with an end of file condition and
handle writes of at least one byte at the maximum file offset by failing
with EFBIG.

Refactor UtilMemoryBuffer to store the file size using off_t instead of
size_t to avoid casts and keep file sizes in the off_t type. Properly
handle errors in the code, such as failing with EROFS instead of EBADF if
the backing memory is not writeable, and failing with EFBIG if writing
beyond the end of the file.

Fix mkpartition not rejecting invalid partition start offsets and lengths.
Strictly enforce partition start and length checks in the partition code.
Enforce partitions exist within regular files or block devices.

Fix a few indention issues.

+++ kernel/fs/util.h
@@ -28,7 +28,7 @@ class UtilMemoryBuffer : public AbstractInode
 {
 public:
 	UtilMemoryBuffer(dev_t dev, ino_t ino, uid_t owner, gid_t group,
-	                 mode_t mode, uint8_t* buf, size_t bufsize,
+	                 mode_t mode, uint8_t* buf, off_t bufsize,
 	                 bool write = true, bool deletebuf = true);
 	virtual ~UtilMemoryBuffer();
 	virtual int truncate(ioctx_t* ctx, off_t length);
@@ -41,7 +41,7 @@ public:
 private:
 	kthread_mutex_t filelock;
 	uint8_t* buf;
-	size_t bufsize;
+	off_t bufsize;
 	bool write;
 	bool deletebuf;
